+++ packages/core/tests/Console.test.ts
@@ -1,8 +1,11 @@
+import exit from 'exit';
 import cliSize from 'term-size';
 import ansiEscapes from 'ansi-escapes';
 import Console from '../src/Console';
 import Output from '../src/Output';
 import { createTestTool } from './helpers';
+
+jest.mock('exit');
 
 describe('Console', () => {
   let cli: Console;
@@ -147,7 +150,13 @@ describe('Console', () => {
       cli.stop = jest.fn();
       cli.handleFailure(error);
 
-      expect(cli.stop).toHaveBeenCalledWith(error, true);
+      expect(cli.stop).toHaveBeenCalledWith(error);
+    });
+
+    it('exits with a code of 2', () => {
+      cli.handleFailure(new Error('Oops'));
+
+      expect(exit).toHaveBeenCalledWith(2);
     });
   });
 
@@ -156,7 +165,13 @@ describe('Console', () => {
       cli.stop = jest.fn();
       cli.handleSignal();
 
-      expect(cli.stop).toHaveBeenCalledWith('Process has been terminated.', true);
+      expect(cli.stop).toHaveBeenCalledWith(new Error('Process has been terminated.'));
+    });
+
+    it('exits with a code of 2', () => {
+      cli.handleSignal();
+
+      expect(exit).toHaveBeenCalledWith(2);
     });
   });
 
@@ -170,6 +185,7 @@ describe('Console', () => {
     it('sets restore flag', () => {
       cli.hideCursor();
 
+      // @ts-ignore Allow access
       expect(cli.restoreCursorOnExit).toBe(true);
     });
 
@@ -177,6 +193,7 @@ describe('Console', () => {
       cli.tool.config.silent = true;
       cli.hideCursor();
 
+      // @ts-ignore Allow access
       expect(cli.restoreCursorOnExit).toBe(false);
     });
   });
@@ -374,6 +391,14 @@ describe('Console', () => {
 
       expect(spy).not.toHaveBeenCalled();
     });
+
+    it('sets `final` flag', () => {
+      expect(cli.isFinalRender()).toBe(false);
+
+      cli.renderFinalOutput(null);
+
+      expect(cli.isFinalRender()).toBe(true);
+    });
   });
 
   describe('resetCursor()', () => {
@@ -394,16 +419,19 @@ describe('Console', () => {
     it('sets restore flag to off', () => {
       cli.showCursor();
 
+      // @ts-ignore Allow access
       expect(cli.restoreCursorOnExit).toBe(false);
     });
   });
 
   describe('start()', () => {
     it('sets `started` flag', () => {
+      // @ts-ignore Allow access
       expect(cli.started).toBe(false);
 
       cli.start();
 
+      // @ts-ignore Allow access
       expect(cli.started).toBe(true);
     });
 
@@ -450,10 +478,12 @@ describe('Console', () => {
     });
 
     it('sets a timer', () => {
+      // @ts-ignore Allow access
       expect(cli.renderTimer).toBeNull();
 
       cli.startRenderLoop();
 
+      // @ts-ignore Allow access
       expect(cli.renderTimer).not.toBeNull();
     });
 
@@ -461,31 +491,37 @@ describe('Console', () => {
       cli.flushOutputQueue = jest.fn();
       cli.startRenderLoop();
 
+      // @ts-ignore Allow access
       const id = cli.renderTimer;
 
       jest.advanceTimersByTime(100);
 
       expect(cli.flushOutputQueue).toHaveBeenCalled();
+      // @ts-ignore Allow access
       expect(cli.renderTimer).not.toEqual(id);
     });
   });
 
   describe('stop()', () => {
     it('sets `stopped` flag', () => {
+      // @ts-ignore Allow access
       expect(cli.stopped).toBe(false);
 
       cli.stop();
 
+      // @ts-ignore Allow access
       expect(cli.stopped).toBe(true);
     });
 
     it('sets `started` flag', () => {
       cli.start();
 
+      // @ts-ignore Allow access
       expect(cli.started).toBe(true);
 
       cli.stop();
 
+      // @ts-ignore Allow access
       expect(cli.started).toBe(false);
     });
 
@@ -505,15 +541,6 @@ describe('Console', () => {
       cli.stop();
 
       expect(spy).toHaveBeenCalledWith(null);
-    });
-
-    it('calls `stop` with string', () => {
-      const spy = jest.fn();
-
-      cli.on('stop', spy);
-      cli.stop('Oops');
-
-      expect(spy).toHaveBeenCalledWith(new Error('Oops'));
     });
 
     it('calls `stop` with error', () => {
@@ -537,7 +564,7 @@ describe('Console', () => {
     it('renders final output with an error', () => {
       const spy = jest.spyOn(cli, 'renderFinalOutput');
 
-      cli.stop('Oops');
+      cli.stop(new Error('Oops'));
 
       expect(spy).toHaveBeenCalledWith(new Error('Oops'));
     });
@@ -548,12 +575,6 @@ describe('Console', () => {
       cli.stop();
 
       expect(spy).toHaveBeenCalledTimes(1);
-    });
-
-    it('throws the error if forced', () => {
-      expect(() => {
-        cli.stop('Thrown', true);
-      }).toThrowError('Thrown');
     });
 
     it('copies logs to error logs if persist is true', () => {
@@ -585,6 +606,7 @@ describe('Console', () => {
       cli.renderTimer = 123;
       cli.stopRenderLoop();
 
+      // @ts-ignore Allow access
       expect(cli.renderTimer).toBeNull();
     });
   });